refactor: extract highlighter into two components
